chores: tsfy all features examples
