+++ examples/demos/components/Formatters/ImageFormatter.tsx
