+++ examples/demos/components/Toolbar/toolbar.less
@@ -1,0 +1,62 @@
+.react-grid-Toolbar {
+  background-color: #ffffff;
+  border-color: #e7eaec;
+  border-image: none;
+  border-style: solid solid none;
+  border-width: 1px 1px 1px 1px;
+  color: inherit;
+  margin-bottom: 0;
+  padding: 14px 15px 7px;
+  height: 48px;
+}
+
+.react-grid-Toolbar .btn {
+  display: inline-block;
+  padding: 6px 12px;
+  margin-bottom: 0;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 1;
+  text-align: center;
+  white-space: nowrap;
+  vertical-align: middle;
+  -ms-touch-action: manipulation;
+  touch-action: manipulation;
+  cursor: pointer;
+  -webkit-user-select: none;
+  color: inherit;
+  background: white;
+  border: 1px solid #e7eaec;
+}
+
+.react-grid-Toolbar .btn:hover {
+  color: inherit;
+  border: 1px solid #d2d2d2;
+}
+
+.react-grid-Toolbar .grouped-col-btn {
+  background-color: #428bca;
+  color: white;
+  border-color: #2b669a;
+}
+
+.react-grid-Toolbar .grouped-col-btn:hover {
+  color: white;
+  border-color: #2b669a;
+}
+
+.react-grid-Toolbar .grouped-col-btn + .grouped-col-btn {
+  margin-left: 5px;
+}
+
+.grouped-col-btn > .MuiSvgIcon-root {
+  font-size: 18px;
+  vertical-align: middle;
+}
+
+.react-grid-Toolbar .tools {
+  float: right;
+  position: relative;
+  padding: 0;
+  margin-top: -6px;
+}
