+++ examples/index.tsx
@@ -11,10 +11,8 @@ import Basic from './demos/example01-basic';
 // import SortableCols from './demos/example08-sortable-cols';
 // import FilterableGrid from './demos/example09-filterable-grid';
 // import OneMillionRows from './demos/example10-one-million-rows';
-// import ImmutableData from './demos/example11-immutable-data';
 // import CustomRowRenderer from './demos/example12-customRowRenderer';
-// import AllFeatures from './demos/example13-all-features';
-// import AllFeaturesImmutable from './demos/example14-all-features-immutable';
+import AllFeatures from './demos/example13-all-features';
 // import EmptyRows from './demos/example15-empty-rows';
 // import CellDragDown from './demos/example16-cell-drag-down';
 // import FilterableSortableGrid from './demos/example16-filterable-sortable-grid';
@@ -26,7 +24,6 @@ import Basic from './demos/example01-basic';
 // import CellSelectionEvents from './demos/example21-cell-selection-events';
 // import Grouping from './demos/example21-grouping';
 // import CustomFilters from './demos/example22-custom-filters';
-// import ImmutableDataGrouping from './demos/example23-immutable-data-grouping';
 // import DraggableHeader from './demos/example24-draggable-header';
 // import TreeView from './demos/example25-tree-view';
 // import TreeViewNoAddDelete from './demos/example26-tree-view-no-add-delete';
@@ -42,18 +39,16 @@ ReactDOM.render((
     <Router>
       <ul className="nav">
         <li><NavLink to="/examples/basic">Basic Example</NavLink></li>
-        <li><NavLink to="/examples/resizable-cols">Resizable Cols Example</NavLink></li>
+        {/* <li><NavLink to="/examples/resizable-cols">Resizable Cols Example</NavLink></li>
         <li><NavLink to="/examples/frozen-cols">Frozen Cols Example</NavLink></li>
         <li><NavLink to="/examples/editable">Editable Example</NavLink></li>
         <li><NavLink to="/examples/custom-formatters">Custom Formatters Example</NavLink></li>
         <li><NavLink to="/examples/built-in-editors">Built In Editors Example</NavLink></li>
         <li><NavLink to="/examples/sortable-cols">Sortable Cols Example</NavLink></li>
         <li><NavLink to="/examples/filterable-grid">Filterable Grid Example</NavLink></li>
-        <li><NavLink to="/examples/immutable-data">Immutable Data Example</NavLink></li>
-        <li><NavLink to="/examples/customRowRenderer">Custom Row Renderer Example</NavLink></li>
+        <li><NavLink to="/examples/customRowRenderer">Custom Row Renderer Example</NavLink></li> */}
         <li><NavLink to="/examples/all-features">All Features Example</NavLink></li>
-        <li><NavLink to="/examples/all-features-immutable">All Features Immutable Example</NavLink></li>
-        <li><NavLink to="/examples/empty-rows">Empty Rows Example</NavLink></li>
+        {/*<li><NavLink to="/examples/empty-rows">Empty Rows Example</NavLink></li>
         <li><NavLink to="/examples/cell-drag-down">Cell Drag Down Example</NavLink></li>
         <li><NavLink to="/examples/filterable-sortable-grid">Filterable Sortable Grid Example</NavLink></li>
         <li><NavLink to="/examples/row-select">Row Select Example</NavLink></li>
@@ -64,7 +59,6 @@ ReactDOM.render((
         <li><NavLink to="/examples/cell-selection-events">Cell Selection Events</NavLink></li>
         <li><NavLink to="/examples/grouping">Grouping Example</NavLink></li>
         <li><NavLink to="/examples/custom-filters">Custom Filters Example</NavLink></li>
-        <li><NavLink to="/examples/immutable-data-grouping">Immutable Data Grouping Example</NavLink></li>
         <li><NavLink to="/examples/draggable-header">Draggable Header Example</NavLink></li>
         <li><NavLink to="/examples/tree-view">Tree View Example</NavLink></li>
         <li><NavLink to="/examples/tree-view-no-add-delete">Tree View No Add Delete Example</NavLink></li>
@@ -73,7 +67,7 @@ ReactDOM.render((
         <li><NavLink to="/examples/descendingFirstSortable">Descending First Sortable Example</NavLink></li>
         <li><NavLink to="/examples/selection-range-events">Selection Range Events Example</NavLink></li>
         <li><NavLink to="/examples/isScrolling">IsScrolling Example</NavLink></li>
-        <li><NavLink to="/examples/summary-rows">Summary Rows Example</NavLink></li>
+        <li><NavLink to="/examples/summary-rows">Summary Rows Example</NavLink></li> */}
       </ul>
       <Route path="/examples/basic">
         <Basic />
@@ -86,10 +80,8 @@ ReactDOM.render((
       {/* <Route path="/examples/sortable-cols" component={SortableCols} /> */}
       {/* <Route path="/examples/filterable-grid" component={FilterableGrid} /> */}
       {/* <Route path="/examples/one-million-rows" component={OneMillionRows} /> */}
-      {/* <Route path="/examples/immutable-data" component={ImmutableData} /> */}
       {/* <Route path="/examples/customRowRenderer" component={CustomRowRenderer} /> */}
-      {/* <Route path="/examples/all-features" component={AllFeatures} /> */}
-      {/* <Route path="/examples/all-features-immutable" component={AllFeaturesImmutable} /> */}
+      <Route path="/examples/all-features" component={AllFeatures} />
       {/* <Route path="/examples/empty-rows" component={EmptyRows} /> */}
       {/* <Route path="/examples/cell-drag-down" component={CellDragDown} /> */}
       {/* <Route path="/examples/filterable-sortable-grid" component={FilterableSortableGrid} /> */}
@@ -101,7 +93,6 @@ ReactDOM.render((
       {/* <Route path="/examples/cell-selection-events" component={CellSelectionEvents} /> */}
       {/* <Route path="/examples/grouping" component={Grouping} /> */}
       {/* <Route path="/examples/custom-filters" component={CustomFilters} /> */}
-      {/* <Route path="/examples/immutable-data-grouping" component={ImmutableDataGrouping} /> */}
       {/* <Route path="/examples/draggable-header" component={DraggableHeader} /> */}
       {/* <Route path="/examples/tree-view" component={TreeView} /> */}
       {/* <Route path="/examples/tree-view-no-add-delete" component={TreeViewNoAddDelete} /> */}
