+++ package.json
@@ -46,11 +46,13 @@
     "@types/classnames": "^2.2.9",
     "@types/enzyme": "^3.10.3",
     "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/faker": "^4.1.8",
     "@types/jest": "^24.0.23",
     "@types/react": "^16.9.14",
     "@types/react-dom": "^16.9.4",
     "@types/react-is": "^16.7.1",
     "@types/react-router-dom": "^5.1.3",
+    "@types/react-virtualized": "^9.21.7",
     "@typescript-eslint/eslint-plugin": "^2.10.0",
     "@typescript-eslint/parser": "^2.10.0",
     "clean-css": "^4.2.1",
