--- packages/react-data-grid-addons/src/editors/index.ts
@@ -1,2 +1,0 @@
-export { SimpleTextEditor } from 'react-data-grid';
-export { default as DropDownEditor } from './DropDownEditor';
