+++ packages/react-data-grid-addons/src/formatters/index.ts
@@ -1,2 +1,1 @@
-export { default as ImageFormatter } from './ImageFormatter';
 export { default as DropDownFormatter } from './DropDownFormatter';
