+++ packages/react-data-grid-addons/src/index.ts
@@ -1,11 +1,9 @@
-import * as Editors from './editors';
-import * as Formatters from './formatters';
-import Toolbar from './toolbars/Toolbar';
-import * as ToolsPanel from './toolbars';
+import * as Filters from './cells/headerCells/filters';
 import * as Data from './data';
-import * as Menu from './menu';
 import * as Draggable from './draggable';
 import * as DraggableHeader from './draggable-header';
-import * as Filters from './cells/headerCells/filters';
+import * as Formatters from './formatters';
+import * as Menu from './menu';
+import * as ToolsPanel from './toolbars';
 
-export { Editors, Formatters, Toolbar, Menu, Data, ToolsPanel, Draggable, DraggableHeader, Filters };
+export { Formatters, Menu, Data, ToolsPanel, Draggable, DraggableHeader, Filters };
