+++ tsconfig.json
@@ -20,6 +20,7 @@
     "types": ["jest"]
   },
   "files": [
-    "src/index.ts"
+    "src/index.ts",
+    "examples/index.tsx"
   ]
 }
